fix: new synced_data_file_logger.py
